requirements.txt
pandas==2.2.2
numpy==2.0.1
faker==26.0.0
psycopg2-binary==2.9.9
pyarrow==17.0.0
sqlalchemy==2.0.32
streamlit==1.37.1
python-dotenv==1.0.1